pyproject.toml
[project]
name = "gke-anthos-hackhathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "fastmcp>=2.12.3",
    "google-adk==1.14.1",
    "psycopg2>=2.9.10",
    "pyjwt>=2.10.1",
    "uvicorn>=0.35.0",
]
